Simplify caching

+++ run.py
@@ -2,7 +2,8 @@ from pprint import pprint
 
 import const
 from search_services.hh.search_service import HeadHunterSearchService
-from search_services.hh.request_consts import URL, HEADERS, PARAMS
+from search_services.request_consts import HH_REQUEST_CONST
+from search_services.storing_const import SEARCHED_TEXT_KEY
 from mongo.services import DAOVacancies, DAOSearchedText
 
 
@@ -12,16 +13,16 @@ def main():
     text = searched_text.strip()
     file_prefix_name = text.replace(' ', '_')
 
-    search_object = HeadHunterSearchService(URL, HEADERS, PARAMS)
+    search_object = HeadHunterSearchService(**HH_REQUEST_CONST)
 
-    vacancies_list = search_object.make_fully_hh_search(text)
+    vacancies_list = search_object.make_hh_searching(text)
     last_searched_data = search_object.get_last_searched_text()
 
     vacancies_collection = DAOVacancies(f'collection_{file_prefix_name}_vacancies')
     repeated_vacancies = []
     if last_searched_data and vacancies_list:
-        searched_text_collection = DAOSearchedText(const.SEARCHED_COLLECTION)
-        searched_text_collection.use_index(const.SEARCHED_TEXT_KEY, unique=True)
+        searched_text_collection = DAOSearchedText(const.SEARCHED_TEXTS_COLLECTION)
+        searched_text_collection.use_index(SEARCHED_TEXT_KEY, unique=True)
         is_new_search = searched_text_collection.insert(last_searched_data)
         if is_new_search:
             vacancies_collection.use_index(const.LINK, unique=True)
@@ -31,11 +32,17 @@ def main():
                                       if i in unsuccessful_indexes]
         else:
             repeated_vacancies = vacancies_list
-        updated_indexes = vacancies_collection.update_many_by_field(repeated_vacancies, const.LINK)
+        updated_indexes, inserted_indexes = vacancies_collection.update_many_by_field(repeated_vacancies, const.LINK)
+        if updated_indexes is None and inserted_indexes is None:
+            print('Nothing has been updated or inserted')
         if updated_indexes:
             print(f'Some vacancies was updated, on indexes {updated_indexes}')
+        if inserted_indexes:
+            print(f'Some vacancies was updated, on indexes {inserted_indexes}')
+
     else:
-        vacancies_collection.drop()
+        if vacancies_collection.is_empty():
+            vacancies_collection.drop()
 
     if vacancies_collection.is_exist():
         filters = {
